+++ src/main/java/me/twister915/punishments/model/PunishmentFactory.java
@@ -22,6 +22,6 @@ import org.bukkit.OfflinePlayer;
 import java.util.Date;
 
 public interface PunishmentFactory<T extends Punishment> {
-    T createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Integer lengthInSeconds);
+    T createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Double lengthInSeconds);
     PunishmentManager<T> getNewManager(BaseStorage<T> storage);
 }
