+++ src/main/java/me/twister915/punishments/model/factory/TemporaryBanFactory.java
@@ -27,7 +27,7 @@ import java.util.Date;
 
 public final class TemporaryBanFactory implements PunishmentFactory<TemporaryBan> {
     @Override
-    public TemporaryBan createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Integer lengthInSeconds) {
+    public TemporaryBan createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Double lengthInSeconds) {
         return new TemporaryBan(active, reason, date, punisherId, player, lengthInSeconds);
     }
