+++ src/main/java/me/twister915/punishments/model/factory/TemporaryMuteFactory.java
@@ -27,7 +27,7 @@ import java.util.Date;
 
 public final class TemporaryMuteFactory implements PunishmentFactory<TemporaryMute> {
     @Override
-    public TemporaryMute createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Integer lengthInSeconds) {
+    public TemporaryMute createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Double lengthInSeconds) {
         return new TemporaryMute(active, reason, date, punisherId, player, lengthInSeconds);
     }
