+++ src/main/java/me/twister915/punishments/model/factory/WarningFactory.java
@@ -27,7 +27,7 @@ import java.util.Date;
 
 public final class WarningFactory implements PunishmentFactory<Warning> {
     @Override
-    public Warning createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Integer lengthInSeconds) {
+    public Warning createNew(OfflinePlayer player, String reason, String punisherId, Date date, boolean active, Double lengthInSeconds) {
         return new Warning(reason, date, punisherId, player);
     }
